ChatGraphics feedback
One highly recommended practice in programming is to declare variables as close to their use as possible. This means that if at all possible, you should avoid using "fields" or "instance variables" for elements that are local scope *in usage*. That means even if it's something that belongs to the class as a whole in terms of actual implementation, if I'm only using the variable locally and I don't have to otherwise declare it in the classes scope, I should still only declare it locally. A good example of this is your JPanel and JScrollPane, which could both be added locally in the "initializeComponents" method and are never used again. A counterpoint of course is the inputText and outputText fields, which have to be used in other methods than only initializeComponents.
The last case which I will remark on here is the use of your guiFont, which you declare outside the method and use inside of it. This is actually a good example of a constant, most likely, where we'd set the font and consider it good. This is typically implemented using "private static Font GUI_FONT = new Font(...)", where the UPPER_SNAKE_CASE indicates you're using a constant.